1. Переписал экран обучения на Clean-swift 2. Координатор переходов для старта приложения 3. Поправил AppDelegate
