+++ MyPets/MyPetsUnitTests/Mock/MockNetworkService/MockURLSession.swift
@@ -8,15 +8,16 @@
 import Foundation
 
 class MockURLSession: Networking {
-    var networkDataTaskDoubleToReturn: NetworkDataTask?
 
-    var completion: (Data?, URLResponse?, Error?)?
+	var networkDataTaskDoubleToReturn: NetworkDataTask?
 
-    func performDataTask(with request: URLRequest,
-                         completionHandler: @escaping (Data?, URLResponse?, Error?) -> Void) -> NetworkDataTask {
-        if let completion = completion {
-            completionHandler(completion.0, completion.1, completion.2)
-        }
-        return networkDataTaskDoubleToReturn ?? NetworkDataTaskSpy()
-    }
+	var completion: (Data?, URLResponse?, Error?)?
+
+	func performDataTask(with request: URLRequest,
+						 completionHandler: @escaping (Data?, URLResponse?, Error?) -> Void) -> NetworkDataTask {
+		if let completion = completion {
+			completionHandler(completion.0, completion.1, completion.2)
+		}
+		return networkDataTaskDoubleToReturn ?? NetworkDataTaskSpy()
+	}
 }
